Card test 80%

+++ src/main/java/com/project/credit/card/service/CreditCardServiceImpl.java
@@ -9,6 +9,7 @@ import com.project.credit.card.exception.CardException;
 import com.project.credit.card.exception.CreditCardRequestException;
 import com.project.credit.card.repository.CreditCardRepository;
 import com.project.credit.card.repository.CreditCardRequestRepository;
+import org.springframework.data.jpa.repository.support.SimpleJpaRepository;
 import org.springframework.stereotype.Service;
 import com.project.credit.customer.entity.Customer;
 import com.project.credit.customer.exception.CustomerException;
@@ -29,6 +30,7 @@ public class CreditCardServiceImpl implements CreditCardService {
     private CreditCardRepository creditCardRepository;
     @Autowired
     private CustomerService customerService;
+
     @Autowired
     private CreditCardRequestRepository creditCardRequestRepository;
 
@@ -168,21 +170,48 @@ public class CreditCardServiceImpl implements CreditCardService {
         return creditCards;
     }
 
+
     @Override
     public CreditCard deleteCreditCardByCardNumber(String cardNumber) throws CardException, CustomerException {
+        if (cardNumber == null || cardNumber.isEmpty()) {
+            throw new CardException("Card number cannot be null or empty");
+        }
         CreditCard creditCard = findCreditCardByCardNumber(cardNumber);
+        if (creditCard == null) {
+            throw new CardException("Credit card not found in database");
+        }
         creditCard.getCustomer().setCreditCard(null);
         customerService.updateCustomer(creditCard.getCustomer());
         creditCardRepository.delete(creditCard);
         return creditCard;
     }
 
-
     @Override
+    public CreditCardRequest deleteCreditCardRequestById(Long creditCardRequestId) throws CreditCardRequestException {
+        if (creditCardRequestId == null) {
+            throw new CreditCardRequestException("Credit card request ID cannot be null");
+        }
+        CreditCardRequest creditCardRequest = getCreditCardRequestById(creditCardRequestId);
+        creditCardRequestRepository.delete(creditCardRequest);
+        return null;
+    }
+    @Override
+    public CreditCardRequest getCreditCardRequestById(Long creditCardRequestId) throws CreditCardRequestException {
+        CreditCardRequest creditCardRequest=creditCardRequestRepository.findById(creditCardRequestId).orElse(null);
+        if(creditCardRequest==null)
+            throw new CreditCardRequestException("No such request found");
+        return creditCardRequest;
+    }
+
+
     public CreditCard updateCreditCard(CreditCard creditCard) throws CardException {
+        if (creditCard == null) {
+            throw new CardException("Credit card cannot be null");
+        }
         if (findCreditCardByCardNumber(creditCard.getCardNumber()) == null)
             throw new CardException("No such credit card found in database");
         return creditCardRepository.save(creditCard);
     }
+
 
 }
